Move StopState & key (most basic structs)
to pathspec.h
